# micrometeorite-openapi-model/swagger/Micrometeorite_SwaggerV1.yaml
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: "Micrometeorite evaluation system"
  contact:
    email: "maximiliangruber91@gmail.com"
  license:
    name: "Apache 2.0"
    url: "https://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/micrometeorite/api/v1"
schemes:
  - "https"
  - "http"
paths:
  /micrometeoriteFind:
    post:
      tags:
        - "micrometeoriteFind"
      summary: "Add micrometeorite find"
      description: "Add micrometeorite find"
      operationId: "addMicrometeoriteFind"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "MicrometeoriteFind object that needs to be added"
          required: true
          schema:
            $ref: "#/definitions/micrometeoriteFind"
      responses:
        "201":
          description: "Micrometeorite Finding created and predicted"
          schema:
            type: array
            items:
              $ref: '#/definitions/predictionResult'
        "405":
          description: "Invalid input"
    get:
      tags:
        - "micrometeoriteFind"
      summary: "Returns all micrometeorite Findings"
      operationId: "getAllMicrometeoriteFinds"
      produces:
        - "application/json"
      responses:
        "200":
          description: "Return all Micrometeorite Findings"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/micrometeoriteFind"

  /micrometeoriteFind/{micrometeoriteFindId}:
    get:
      tags:
        - "micrometeoriteFindById"
      summary: "Returns micrometeorite Finding by ID"
      operationId: "getMicrometeoriteFindById"
      produces:
        - "application/json"
      parameters:
        - name: micrometeoriteFindId
          in: "path"
          description: "micrometeoriteFindId"
          required: true
          type: "integer"
      responses:
        "200":
          description: "Return Micrometeorite Finding by ID"
          schema:
            $ref: "#/definitions/micrometeoriteFind"
    delete:
      tags:
        - "micrometeoriteFind"
      summary: "Deletes a MicrometeoriteFind"
      description: ""
      operationId: "deleteMicrometeoriteFind"
      produces:
        - "application/json"
      parameters:
        - name: micrometeoriteFindId
          in: "path"
          description: "micrometeoriteFindId id to delete"
          required: true
          type: "integer"
      responses:
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "MicrometeoriteFind not found"

definitions:
  micrometeoriteFind:
    type: "object"
    properties:
      micrometeoriteFindId:
        type: "number"
        format: "int"
      micrometeoriteFindDate:
        type: "string"
        format: "date-time"
      micrometeoriteFindPlace:
        type: "string"
      micrometeoriteFindPlaceDescription:
        type: "string"
      micrometeoriteFindCoordinates:
        type: "number"
        format: "double"
      micrometeoriteDiameter:
        type: "number"
        format: "double"
      micrometeoriteForm:
        type: "string"
      micrometeoriteWeight:
        type: "number"
        format: "double"
      micrometeoriteChemicalComposition:
        type: "string"
      micrometeoriteFindComment:
        type: "string"
      micrometeoriteFindFinder:
        $ref: "#/definitions/person"
      micrometeoriteFindRecorder:
        $ref: "#/definitions/person"
      images: # <-- image embedded into JSON
        type: "array"
        items:
          $ref: "#/definitions/image"

  image:
    type: "object"
    properties:
      imageId:
        type: "number"
        format: "int"
      photographyDate:
        type: "string"
        format: "date-time"
      photographer:
        $ref: "#/definitions/person"
      recordingInstrument:
        type: "string"
      magnification:
        type: "string"
      camera:
        type: "string"
      lens:
        type: "string"
      micrometeoritePredictionModelName:
        type: "string"
      micrometeoritePrediction:
        type: "number"
        format: "int"
      picture: # <-- image embedded into JSON
        type: "string"
        format: "byte"
        description: "Base64-encoded contents of the avatar image"

  person:
    type: "object"
    properties:
      name:
        type: "string"
      firstname:
        type: "string"
      email:
        type: "string"
      website:
        type: "string"
      birthday:
        type: "string"
        format: "date"
      street:
        type: "string"
      zipcode:
        type: integer
      location:
        type: "string"
      country:
        type: "string"
      phonenumber:
        type: "string"

  predictionResult:
    type: "object"
    properties:
      micrometeoriteFindId:
        type: "number"
        format: "int"
      imagesResults: # <-- image embedded into JSON
        type: "array"
        items:
          $ref: "#/definitions/imageResult"

  imageResult:
    type: "object"
    properties:
      imageId:
        type: "number"
        format: "int"
      micrometeoritePredictionModelName:
        type: "string"
      micrometeoritePrediction:
        type: "number"
        format: "int"